Create README.md
In this capstone project, I made typographic and color design choices for Colmar Academy, and build their responsive site from scratch
I was  tasked with designing and creating the landing page for a new school. I have implemented my own design. 
For this project I was provided with wireframes. A wireframe is a document which contains the positioning of the elements on a page without specifying any other design information.
For this project I will implement the site specified by the wireframes and make all remaining design decisions to complete the site. You can find the wireframes for this project here and all of the necessary assets for the project here.